molly-scraper/requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
extruct==0.18.0
flexcache==0.3
flexparser==0.4
floret==0.10.5
grpcio==1.71.0
grpcio-tools==1.71.0
html5lib==1.1
html_text==0.7.0
idna==3.10
ingredient_parser_nlp==2.0.0
isodate==0.7.2
joblib==1.4.2
jstyleson==0.0.2
lxml==5.3.1
lxml_html_clean==0.4.1
mf2py==2.0.1
nltk==3.9.1
numpy==2.2.4
pika==1.3.2
Pint==0.24.4
platformdirs==4.3.7
protobuf==5.29.4
pyparsing==3.2.1
pyRdfa3==3.6.4
python-crfsuite==0.9.11
rdflib==7.1.3
recipe_scrapers==15.6.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
w3lib==2.3.1
webencodings==0.5.1
